requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Autils==0.4.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.2
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
configparser==5.2.0
constantly==15.1.0
crypto==1.4.1
cryptography==40.0.2
daphne==3.0.2
Django==4.1.4
django-apscheduler==0.6.2
django-article==0.1.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-crypto==0.2.0
django-filter==21.1
django-js-asset==2.0.0
django-mysql==4.9.0
django-redis==5.2.0
django-simpleui==2022.12.18
django-sslserver==0.22
django-tinymce==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
edit==0.0.4
exceptiongroup==1.1.1
frozenlist==1.3.3
future==0.18.2
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.11.3
incremental==22.10.0
it-drf-utils==0.2.3
kombu==5.2.4
lxml==4.9.2
Markdown==3.3.6
msgpack==1.0.5
multidict==6.0.4
mysqlclient==2.1.1
Naked==0.1.32
numpy==1.24.0
openai==0.27.4
outcome==1.2.0
pefile==2022.5.30
Pillow==9.5.0
prompt-toolkit==3.0.36
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyJWT==2.6.0
PyMySQL==1.0.2
pyOpenSSL==23.1.1
pyotp==2.8.0
PySocks==1.7.1
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
qiniu==7.10.0
redis==4.5.5
requests==2.28.2
selenium==4.9.0
service-identity==21.1.0
shellescape==3.8.1
simpleui==4.0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.5
wsproto==1.2.0
xlwt==1.3.0
yarl==1.9.1
zipp==3.8.0
zope.interface==6.0
